add a new retry strategy
The ISSN site will sometimes return "xml" data, when requesting "json" -
but only sometimes and also, it is hard to reproduce.

The server might reuse a class variable ("output format") in different
instances - but that is just a guess.

Anyway, we retry at most ten times, if we encounter an undecodable
response body. This might help.